+++ src/Paket.Core/InstallProcess.fs
@@ -2,7 +2,7 @@
 module Paket.InstallProcess
 
 open Paket
-open Chessie.Rop
+open Chessie.ErrorHandling
 open Paket.Domain
 open Paket.Logging
 open Paket.BindingRedirects
@@ -135,7 +135,7 @@ let findAllReferencesFiles root =
                                 |> Option.map (fun r -> p, r))
     |> Array.map (fun (project,file) -> 
         try 
-            succeed <| (project, ReferencesFile.FromFile(file))
+            ok <| (project, ReferencesFile.FromFile(file))
         with _ -> 
             fail <| ReferencesFileParseError (FileInfo(file)))
     |> collect