+++ src/Paket.Core/PackageProcess.fs
@@ -8,7 +8,7 @@ open Paket.Domain
 open Paket.Logging
 open System.Collections.Generic
 open Paket.PackageMetaData
-open Chessie.Rop
+open Chessie.ErrorHandling
 
 let Pack(dependencies : DependenciesFile, packageOutputPath, buildConfig, version, releaseNotes) =
     let buildConfig = defaultArg buildConfig "Release"  
@@ -118,7 +118,7 @@ let Pack(dependencies : DependenciesFile, packageOutputPath, buildConfig, versio
                     NupkgWriter.Write core optional (Path.GetDirectoryName templateFile.FileName) packageOutputPath
                     verbosefn "Packed: %s" templateFile.FileName
                 | IncompleteTemplate -> 
-                    failwithf "There was an attempt to pack incomplete template file %s." templateFile.FileName
+                    failwithf "There was an trial to pack incomplete template file %s." templateFile.FileName
             })
     |> Async.Parallel
     |> Async.RunSynchronously